Introduced W2W support
This update introduces a Boolean IID variable that allows the Mod DLL to select W2W (Wall To Wall) buildings to grow or not.

If Occupant Group contains the boolean variable (0xD02C802E) and the W2W button has been selected then the game will only grow W2W buildings for all zones.

The modder will need to remove this variable for buildings that are not W2W