Building Styles: Add a property to identify PIMX
The absence of this property allows the DLL to detect the older PIMX
versions that accidentally used the 0x2004 community style id as a
placeholder.

This allows those older versions to be treated as using the Maxis
styles, while still allowing for a future community style to use
the 0x2004 value.