3.0 Release (#12)
* AMPS Update

* Dummy Building Style & CAM 3.0

* Fix Typo

* Fix Building Style

* Casino Specialization. Bugfixes

* Add the new Custom Ordinance Host DLL properties (#8)

* Add the new Custom Ordinance Host DLL properties

* Add the residential ordinance demand effects

* Rename the industrial demand effects to match new_properties.xml

* Change Float to Float32

* Fix Conflict Version

---------

Co-authored-by: ulisse99 <88635118+UlisseWolf@users.noreply.github.com>

* Add Submenu CAM and Incinerator

* Fix Bug

* Fixed Casinò

* Fixes and Balance Part 1

* Add the Regional Supply/Demand DLL properties (#9)

* Add the Regional Supply/Demand DLL properties

* Fix some bugs

---------

Co-authored-by: UlisseWolf <88635118+UlisseWolf@users.noreply.github.com>

* Bugfixes

* Fixed W2W

* Improved W2W Filter

* Balance Part 2

* Add the Building Style Exemplar properties (#10)

* Add the Building Style Exemplar properties

* Fix Bugs

---------

Co-authored-by: UlisseWolf <88635118+UlisseWolf@users.noreply.github.com>

* Building Styles: Add a property to identify PIMX (#11)

* Building Styles: Add a property to identify PIMX

The absence of this property allows the DLL to detect the older PIMX
versions that accidentally used the 0x2004 community style id as a
placeholder.

This allows those older versions to be treated as using the Maxis
styles, while still allowing for a future community style to use
the 0x2004 value.

* Fix Bugs

---------

Co-authored-by: UlisseWolf <88635118+UlisseWolf@users.noreply.github.com>

* Update README.md

* CAMInflation

FINAL balancing of construction, demolition and maintenance costs. If it seems high to you for some values consider that there has been simoleon inflation. This is 2025 not 2004

* Update Exemplar Patching

* Removed Automatic generator Building Style

Because certain people broke my fucking balls I had to remove the generator since PIM-X (Shitty Program) can't generate empty prorities.

* Update Power Menu

* AMPS Update

This update modifies the energy part by including operations to make lots compatible with CAM and AMPS.

---------

Co-authored-by: Nicholas Hayes <0xC0000054@users.noreply.github.com>